requirements.txt
beautifulsoup4==4.12.3
black==24.8.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.15.4
idna==3.8
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pbr==6.1.0
platformdirs==4.2.2
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
stevedore==5.3.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
